change int to size_t

+++ ft_strlen.c
@@ -9,10 +9,11 @@
 /*   Updated: 2025/05/02 00:29:08 by lucpardo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-int	ft_strlen(const char	*str)
+size_t	ft_strlen(const char	*str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
